* model/primitive.hh (unsigned_right_shift): Use long long. 	* model/switch.cc (resolve): Use better error locations.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gcjx-branch@108828 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcjx/model/primitive.hh
@@ -458,8 +458,8 @@ public:
   {
     // Only 'long' and 'int' shifts are supported.
     assert (sig_char == 'I' || sig_char == 'J');
-    unsigned long lval = (unsigned long) (T) l;
-    unsigned long rval = (unsigned long) (T) r;
+    unsigned long long lval = (unsigned long long) (T) l;
+    unsigned long long rval = (unsigned long long) (T) r;
     // This is a little ugly -- there should be a way to compute the
     // mask.
     if (sig_char == 'I')
